pyproject.toml
[project]
name = "local-mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "euriai>=1.0.32",
    "fastmcp>=3.1.0",
    "langchain>=1.2.10",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.10",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
